Transfer AbstractTest to IAbstractTest. Leave AbstractTest for compatibility
