+++ carina-core/src/main/java/com/qaprosoft/carina/core/foundation/AbstractTest.java
@@ -15,89 +15,10 @@
  *******************************************************************************/
 package com.qaprosoft.carina.core.foundation;
 
-import java.lang.annotation.Annotation;
-
-import org.testng.ITestContext;
-import org.testng.ITestNGMethod;
-import org.testng.SkipException;
-import org.testng.annotations.BeforeClass;
-import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.BeforeSuite;
-import org.testng.annotations.DataProvider;
-import org.testng.annotations.Listeners;
-
-import com.nordstrom.automation.testng.LinkedListeners;
-import com.qaprosoft.carina.core.foundation.dataprovider.core.DataProviderFactory;
-import com.qaprosoft.carina.core.foundation.listeners.CarinaListener;
-import com.qaprosoft.carina.core.foundation.report.testrail.ITestCases;
-import com.qaprosoft.carina.core.foundation.utils.Configuration;
-import com.qaprosoft.carina.core.foundation.utils.Configuration.Parameter;
-import com.qaprosoft.carina.core.foundation.utils.common.CommonUtils;
-import com.qaprosoft.carina.core.foundation.utils.factory.ICustomTypePageFactory;
-import com.zebrunner.agent.core.registrar.CurrentTest;
-import com.zebrunner.agent.testng.listener.DataProviderInterceptor;
-import com.zebrunner.agent.testng.listener.TestRunListener;
-
 /*
  * AbstractTest - base test for UI and API tests.
  */
 
-// https://github.com/qaprosoft/carina/issues/951
-// reused com.nordstrom.tools.testng-foundation to register ordered listeners
+public abstract class AbstractTest implements IAbstractTest {
 
-// on start order is TestRunListener and CarinaListener
-// on finish reverse order, i.e. CarinaListener, TestRunListener
-@LinkedListeners({ CarinaListener.class, TestRunListener.class, DataProviderInterceptor.class })
-@Listeners({DataProviderInterceptor.class})
-public abstract class AbstractTest implements ICustomTypePageFactory, ITestCases {
-
-    protected static final long EXPLICIT_TIMEOUT = Configuration.getLong(Parameter.EXPLICIT_TIMEOUT);
-    
-	@BeforeSuite(alwaysRun = true)
-	private void onCarinaBeforeSuite() {
-		// do nothing
-	}
-
-	@BeforeClass(alwaysRun = true)
-	private void onCarinaBeforeClass() {
-		// do nothing
-	}
-
-	@BeforeMethod(alwaysRun = true)
-	private void onCarinaBeforeMethod() {
-		// do nothing
-	}
-
-    @DataProvider(name = "DataProvider", parallel = true)
-    public Object[][] createData(final ITestNGMethod testMethod, ITestContext context) {
-        Annotation[] annotations = testMethod.getConstructorOrMethod().getMethod().getDeclaredAnnotations();
-        Object[][] objects = DataProviderFactory.getDataProvider(annotations, context, testMethod);
-        return objects;
-    }
-
-    @DataProvider(name = "SingleDataProvider")
-    public Object[][] createDataSingleThread(final ITestNGMethod testMethod,
-            ITestContext context) {
-        Annotation[] annotations = testMethod.getConstructorOrMethod().getMethod().getDeclaredAnnotations();
-        Object[][] objects = DataProviderFactory.getDataProvider(annotations, context, testMethod);
-        return objects;
-    }
-    
-    /**
-     * Pause for specified timeout.
-     *
-     * @param timeout in seconds.
-     */
-    public void pause(long timeout) {
-        CommonUtils.pause(timeout);
-    }
-
-    public void pause(Double timeout) {
-        CommonUtils.pause(timeout);
-    }
-    
-    protected void skipExecution(String message) {
-        CurrentTest.revertRegistration();
-        throw new SkipException(message);
-    }
 }
